Added: ActiveSupport as integration, extracted cache instrumentation from Rails.
