+++ lib/ddtrace/contrib/active_support/instrumentation.rb
@@ -1,22 +1,14 @@
-require 'thread'
-require 'ddtrace/contrib/rails/ext'
+require 'ddtrace/contrib/active_support/ext'
 
 module Datadog
   module Contrib
-    module Rails
-      # Code used to create and handle 'rails.cache' spans.
-      module ActiveSupport
-        include Datadog::Patcher
+    module ActiveSupport
+      # Defines instrumentation for ActiveSupport
+      module Instrumentation
+        module_function
 
-        def self.instrument
-          do_once(:instrument) do
-            # patch Rails core components
-            Datadog::RailsCachePatcher.patch_cache_store
-          end
-        end
-
-        def self.start_trace_cache(payload)
-          tracer = Datadog.configuration[:rails][:tracer]
+        def start_trace_cache(payload)
+          tracer = Datadog.configuration[:active_support][:tracer]
 
           # In most of the cases Rails ``fetch()`` and ``read()`` calls are nested.
           # This check ensures that two reads are not nested since they don't provide
@@ -31,7 +23,7 @@ module Datadog
           tracing_context = payload.fetch(:tracing_context)
 
           # create a new ``Span`` and add it to the tracing context
-          service = Datadog.configuration[:rails][:cache_service]
+          service = Datadog.configuration[:active_support][:cache_service]
           type = Ext::SPAN_TYPE_CACHE
           span = tracer.trace(Ext::SPAN_CACHE, service: service, span_type: type)
           span.resource = payload.fetch(:action)
@@ -40,7 +32,7 @@ module Datadog
           Datadog::Tracer.log.debug(e.message)
         end
 
-        def self.finish_trace_cache(payload)
+        def finish_trace_cache(payload)
           # retrieve the tracing context and continue the trace
           tracing_context = payload.fetch(:tracing_context)
           span = tracing_context[:dd_cache_span]
@@ -48,8 +40,11 @@ module Datadog
 
           begin
             # discard parameters from the cache_store configuration
-            store, = *Array.wrap(::Rails.configuration.cache_store).flatten
-            span.set_tag(Ext::TAG_CACHE_BACKEND, store)
+            if defined?(::Rails)
+              store, = *Array.wrap(::Rails.configuration.cache_store).flatten
+              span.set_tag(Ext::TAG_CACHE_BACKEND, store)
+            end
+
             normalized_key = ::ActiveSupport::Cache.expand_cache_key(payload.fetch(:key))
             cache_key = Datadog::Utils.truncate(normalized_key, Ext::QUANTIZE_CACHE_MAX_KEY_SIZE)
             span.set_tag(Ext::TAG_CACHE_KEY, cache_key)
