+++ lib/ddtrace/contrib/rails/ext.rb
@@ -6,17 +6,9 @@ module Datadog
         APP = 'rails'.freeze
         ENV_ANALYTICS_ENABLED = 'DD_RAILS_ANALYTICS_ENABLED'.freeze
         ENV_ANALYTICS_SAMPLE_RATE = 'DD_RAILS_ANALYTICS_SAMPLE_RATE'.freeze
-        QUANTIZE_CACHE_MAX_KEY_SIZE = 300
-        RESOURCE_CACHE_DELETE = 'DELETE'.freeze
-        RESOURCE_CACHE_GET = 'GET'.freeze
-        RESOURCE_CACHE_SET = 'SET'.freeze
         SPAN_ACTION_CONTROLLER = 'rails.action_controller'.freeze
-        SPAN_CACHE = 'rails.cache'.freeze
         SPAN_RENDER_PARTIAL = 'rails.render_partial'.freeze
         SPAN_RENDER_TEMPLATE = 'rails.render_template'.freeze
-        SPAN_TYPE_CACHE = 'cache'.freeze
-        TAG_CACHE_BACKEND = 'rails.cache.backend'.freeze
-        TAG_CACHE_KEY = 'rails.cache.key'.freeze
         TAG_LAYOUT = 'rails.layout'.freeze
         TAG_ROUTE_ACTION = 'rails.route.action'.freeze
         TAG_ROUTE_CONTROLLER = 'rails.route.controller'.freeze
